pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["devmianharoon <devmianharoon>"]
readme = "README.md"


[tool.poetry.scripts]
dev="app.main:start"

[tool.poetry.dependencies]
python = "^3.12"
langgraph = "^0.2.59"
langchain-core = "^0.3.25"
langchain-google-genai = "^2.0.7"
passlib = "^1.7.4"
jwt = "^1.3.1"
python-jose = "^3.3.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
fastapi = "^0.115.6"
sqlmodel = "^0.0.22"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
langchain-community = "^0.3.12"
langchain-text-splitters = "^0.3.3"
langchain = "^0.3.12"
faiss-cpu = "^1.9.0.post1"
ipython = "^8.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
